experiments/requirements.txt
accelerate
diffusers
hydra-core
numpy==1.26.1
imageio
imageio-ffmpeg
ninja
opencv-python-headless
plotly
rich
scipy
timm
torch-ema
tqdm
transformers
wandb
open3d